[gui/Widget] 抽取公共组件-滑条
